requirements.txt
# Core AI/ML
langchain>=0.1.0
langgraph>=0.0.40
langchain-community>=0.0.20
google-generativeai>=0.3.0
pydantic>=2.0.0

# Video/Audio Processing
yt-dlp>=2025.8.0
moviepy>=2.2.1
youtube-transcript-api>=1.2.0
pytube>=15.0.0

# Transcription Fallback (Optional)
openai-whisper>=20231117
torch>=2.0.0

# Utilities
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.0.0
python-json-logger>=2.0.0

# Additional utilities
requests>=2.31.0
pathlib
typing-extensions>=4.0.0
flask>=3.1.0 